Fix csrf token validation issue
Because we're using Caddy to terminate all SSL/TLS certs before getting to Zammad the web application was throwing errors that the CSRF token was invalid. To fix this problem we had to make use of a newly implemented set of environment variables in Zammad to allow the use of trusted proxies and update the default nginx server scheme. To make sure things would work properly I also dropped everything in their own named networks but I don't think this is entirely necessary.

I also had some issues with the indentation so I standardized the indentation for some of the lines as well.